makefoxsrv:
Created FoxWorker class as part of a new worker management framework.

Workers are now configured through the 'workers' DB table, and the program will check here and start connections for all of them on initial startup.

It will also query each worker for the available list of models and store those in 'worker_models'.

Workers will automatically detect when they've gone offline, mark themselves as offline in the DB, and retry periodically.

Added /model command; support for selecting different available models.

Model lists are generated based on which workers are currently connected and what models each one has installed.

Altered queue scheduling function to take model into consideration -- it won't attempt to assign your request to a worker that doesn't have that model available.

Changed namespace from makefoxbot to makefoxsrv, as this will eventually become a general purpose server including websockets and other functions.  The project as a whole is known as makefoxbot.

Web:

Corrected login issues with the statistics pages, as well as some issues with the logout routine and session storage.