Added delay after worker is detected coming back online to prevent a bug where a1111 would crash if you made API requests before it was ready.
Completely overhauled how user settings work; only settings that the user explicitly changes are stored, so that defaults can be applied for the unaltered ones if the defaults change in the future.

Added initial work for FoxSettings class to load defaults and other settings from the database that can be changed/updated easily during runtime.

Added error message if you attempt to generate without a prompt set.